@@ -3,14 +3,14 @@
   "type": "commonjs",
   "scripts": {
     "lint": "xo-yoursunny --fix",
-    "postinstall": "node postinstall.js",
-    "start": "nodemon -w .env -w src server.js",
+    "postinstall": "node postinstall.mjs",
+    "start": "nodemon -w .env -w src src/main.js",
     "typecheck": "tsc"
   },
-  "packageManager": "pnpm@8.1.0",
+  "packageManager": "pnpm@8.6.0",
   "dependencies": {
     "@fastify/formbody": "^7.4.0",
-    "@fastify/static": "^6.9.0",
+    "@fastify/static": "^6.10.2",
     "@fastify/view": "^7.4.1",
     "@ndn/cli-common": "https://ndnts-nightly.ndn.today/cli-common.tgz",
     "@ndn/endpoint": "https://ndnts-nightly.ndn.today/endpoint.tgz",
@@ -24,13 +24,11 @@
     "@ndn/tlv": "https://ndnts-nightly.ndn.today/tlv.tgz",
     "@ndn/util": "https://ndnts-nightly.ndn.today/util.tgz",
     "@strattadb/environment": "^8.0.2",
-    "dotenv": "^16.0.3",
+    "dotenv": "^16.1.3",
     "ejs": "^3.1.9",
     "envfile": "^6.18.0",
     "fast-chunk-string": "^1.0.1",
-    "fastify": "^4.15.0",
-    "graceful-fs": "^4.2.11",
-    "hard-rejection": "^2.1.0",
+    "fastify": "^4.17.0",
     "leveldown": "^6.1.1",
     "nodemon": "^2.0.22",
     "numd": "^3.1.0"
@@ -40,9 +38,9 @@
     "@types/fast-chunk-string": "^1.0.1",
     "@types/graceful-fs": "^4.1.6",
     "@types/leveldown": "^4.0.3",
-    "@types/node": "^18.15.11",
+    "@types/node": "^20.2.5",
     "@typescript/lib-dom": "npm:@types/web@0.0.99",
-    "@yoursunny/xo-config": "0.53.1",
-    "typescript": "~5.0.2"
+    "@yoursunny/xo-config": "0.54.0",
+    "typescript": "~5.1.3"
   }
 }
